requirements.txt
Pillow==10.4.0
colorthief==0.2.1
